Refactored folder hierarchy
