+++ src/index.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom/client';
 
-import App from './App';
+import App from './components/App';
 
 const rootElement = document.getElementById('root');
